feat: add script cautions before performing destructive operations

+++ nvim-install.sh
@@ -1,5 +1,4 @@
 #!/bin/bash
-
 # Exit immediately if a command exits with a non-zero status
 set -e
 
@@ -11,30 +10,37 @@ BIN_DIR="$INSTALL_DIR/bin"
 # Determine the directory where the script is located
 SCRIPT_DIR="$(cd "$(dirname "${BASH_SOURCE[0]}")" && pwd)"
 
-# Download the latest Neovim release
-curl -LO "$NVIM_URL" 
+echo "Downloading Neovim from ${NVIM_URL}..."
+curl -LO "$NVIM_URL"
 
-# Remove any existing Neovim installation
+echo
+echo "WARNING: This script will remove any existing Neovim installations located at /opt/nvim and ${INSTALL_DIR}."
+read -p "Do you wish to continue? [y/N] " confirmation
+if [[ ! "$confirmation" =~ ^[Yy]$ ]]; then
+  echo "Installation cancelled."
+  exit 1
+fi
+
+echo "Removing existing Neovim installations..."
 sudo rm -rf /opt/nvim
 sudo rm -rf "$INSTALL_DIR"
 
-# Extract the downloaded archive to /opt
-sudo tar -C /opt -xzf "$NVIM_ARCHIVE" 
+echo "Extracting the downloaded archive to /opt..."
+sudo tar -C /opt -xzf "$NVIM_ARCHIVE"
 
-# Remove the downloaded archive to clean up
+echo "Cleaning up the downloaded archive..."
 rm "$NVIM_ARCHIVE"
 
 # Add Neovim to the PATH in ~/.bashrc if it's not already there
 if ! grep -q "$BIN_DIR" ~/.bashrc; then
-    echo 'export PATH="$PATH:/opt/nvim-linux64/bin"' >> ~/.bashrc
+    echo "export PATH=\"\$PATH:${BIN_DIR}\"" >> ~/.bashrc
     echo "Added Neovim to PATH in ~/.bashrc"
 else
     echo "Neovim path already exists in ~/.bashrc"
 fi
 
-# Because of how scripts are executed in a subshell, this will not propagate to
-# the parent shell unless its explicitly specified to run the script in the current shell
-# e.g. source <script.sh> or . <script.sh>
+# Source ~/.bashrc to update the PATH for the current session
+# (Note: This may not propagate to the parent shell.)
 source ~/.bashrc
 
 # Verify Neovim installation
@@ -45,16 +51,12 @@ else
     exit 1
 fi
 
-# Move the 'nvim' directory from the script's directory to ~/.config
+# Move the 'nvim' directory from the script's directory to ~/.config if it exists
 NVIM_CONFIG_SOURCE="$SCRIPT_DIR"
 NVIM_CONFIG_DEST="$HOME/.config/nvim"
 
-# Check if the source directory exists
 if [ -d "$NVIM_CONFIG_SOURCE" ]; then
-    # Create the destination directory if it doesn't exist
     mkdir -p "$(dirname "$NVIM_CONFIG_DEST")"
-    
-    # Move the nvim directory
     mv "$NVIM_CONFIG_SOURCE" "$NVIM_CONFIG_DEST"
     echo "Moved 'nvim' configuration to ~/.config/nvim"
 else
